Refactor: simplify `paste_animedit_keys()` function
Instead of a `for` loop for counting, and then a `switch` with a `case`
for each value of the counter to select the right 'matcher' function, just
loop over a vector of matcher functions.

Also I removed some unnecessary casts (the casted value was already
available in a local variable), and reordered some code.

No functional changes.