Fix: Broken usages of `ENUM_OPERATORS()` (not all!)
From what I can tell all these uses of `ENUM_OPERATORS()` are incorrect,
resulting in a wrong bitwise invert result (`~` operator). There are
a bunch of more cases that look incorrect to me, but the ones here are
more clearly wrong, I don't have time to investigate the others
properly.

It seems that `ENUM_OPERATORS()` is fragile in practical use, at the
very least it's unclear how to use it. We need to do this better
somehow.