Fix: `BRUSH_INVERT_TO_SCRAPE_FILL` flag applies for too many brushes
The `BRUSH_INVERT_TO_SCRAPE_FILL` check inside `brush_flip` applies to
any brush instead of just the Fill and Scrape brush, this means that if
an asset is in a weird state where the user has enabled this flag but
switched away from the brush type, the flag still applies even though
the option is no longer visible.

This commit checks the brush type to ensure it only applies to the
specified types.