Refactor: Grease Pencil: Remove kernel function for `vgroup_select_verts`
This was implemented in the kernel but it shouldn't have been.
Since the function operates at the editor level and doesn't
only work at the low-level, it should be in the `object_vgroup.cc` file
in the `editor` context.

This also means that we can use `ensure_selection_attribute` which
actually fixes a crash when the selection attribute is on the wrong
domain.