UI: Experiment: Popup Block Alert Level Indicator
We discussed this a bit in our UI Module meeting.  The idea is to show
a color line at the top of dialogs to indicate the alert level. Orange
for warnings, Blue for informational alerts, Red for errors, etc. This
PR allows any popup block to have an alert level, defaulting to none.