Math: Work around random Windows math::is_negative build failures
Don't use templates for Matrix is_negative, which is randomly tripping
up MSVC.

Note that these already have the assumption of a 3D affine transform
baked in. For 2D or 4D affine transforms the implementation would have
to be different. So templating these for arbitrary dimensions does not
make sense.