Render: Use last render time for remaining time estimate
With adaptive sampling the render time estimate is often very far off, so
instead use this approximation. This means there is no estimate for the
first frame.

This estimation is now computed by the render pipeline rather than the render
engine, so add-ons may need to be updated if they were computing their own.