Fix: Menu search crash after recent PointerRNA change
Caused by 45f231141d55d2936a35fad78960cf861d7b37f5.

The problem was that the now non-trivial PointerRNA type was used inside
of a union. To fix that, replace it with a variant, and generally move
the file towards proper C++ patterns. Also replace the memory arena with
a ResourceScope, of  change from dynstr to the fmt::memory_buffer, and
use a destructor instead of manually freeing memory.