Fix #133943: Unnecessary image full update mark on file open
This would be done when the frame, layer, pass or view changes compared to
the previous value. But for cases like old files without these members or
loading the image datablock into a different scene, this considers the image
to be always changed on file load.

Now always this state on file load, and don't consider the initialize state
as an image update.

This could also happen in the middle of GPU rendering, causing the GPU
texture to be freed while still in use.