Paint: Make brush.asset_activate perform a blocking load
Depending on internal details of how Blender is run, attempting to load
elements from the asset library may either be synchronous & blocking or
asynchronously.

WHen executing a script in background mode, prior to this commit,
operators that are dependent on the asset system will not execute
correctly due to the loading not being complete.

Busy-waiting for this by repeatedly calling the operator over and over
again in python does not resolve. To match behavior of other operators
when called from python scripts such as the quadriflow remesh, this
commit changes the `brush.asset_activate` operator and dependent code to
force a blocking call instead of optionally using the wmJob background
abstraction system.