Refactor: Tooltip Calculation of Sizing and Spacing
Refactor how tooltips are sized, items placed, margins calculated, etc.
The code has some mistakes that we use cludges to compensate for, and
it doesn't properly consider the font size in calculating margins. The
result shown to users in normal circumstances (changing resolution
scale only) should look identical to current. Small changes when the
text style size is changed. However, this is mostly to allow later
adjustments to padding, which currently _breaks_ when changed.