Fix: Weight Paint undo ignores non-tool editor changes
Around when 2.80 was released, 4a08b974f4 was introduced to avoid
pushing undo steps for various changes. The main motivation of the
change was to avoid brush settings polluting the undo history inside
sculpt mode, where the undo steps were not effective.

For Weight Paint, the Sculpt undo system was never used, so to fix this
undesirable behavior, this commit adds a check when inside weight paint
mode to only skip undo steps for settings stored in the `ToolSettings`
struct and inside the Brush datablock.

Ref: #69760 and many others, including #134220 recently

Based on !117417 by Alexander Gavrilov