Fix #132782: MetalRT: Missing Geometry in Cycles preview on MacOS 15.2
The issue also happens on macOS 15.3.

This is a Metal driver bug, a fix is coming in macOS 15.4. Until then
use static BVH for the viewport. There is no perceptible benefit from
refitting, so while it might be less that ideal it allows to side step
the problem and still benefit from the HWRT.

Do it at the sync level to ensure parameters are consistent all over
the place. Alternative is to force it on the Metal BVH level so that
other users of Cycles will get the fix, but it seems to be a bit more
involved change.