Cleanup: Replace `MEM_[cm]allocN` by `MEM_cnew`.
The main issue of alloc-types of allocation is that there is no check on
allocated type possible.

This is a serious source of annoyance (and crashes) when making some
low-level structs non-trivial, as tracking down all usages of these
structs in higher-level other structs and their allocation is... really
painful.

`MEM_cnew` on the other does check that the given type is trivial, at
build time (static assert), which makes such issue... trivial to catch.