MEM_guardedalloc: Refactor to add more type-safety.
The main goal of these changes are to improve static (i.e. build-time)
checks on whether a given data can be allocated and freed with `malloc`
and `free` (C-style), or require proper C++-style construction and
destruction (`new` and `delete`).

* Add new `MEM_malloc_arrayN_aligned` API.
* Add new `MEM_cnew_uninitialized` API to have a type-safe equivalent to
  `MEM_mallocN` in C++ code.
* Add new `MEM_cfree` C++ wrapper around `MEM_freeN`.
* Update `MEM_SAFE_FREE` to use `MEM_cfree` when possible (i.e. not
  handling a `void` pointer).
* Add `MEM_SAFE_DELETE`, similar to `MEM_SAFE_FREE` but calling
  `MEM_delete`.

The changes to `MEM_SAFE_FREE` required a few removal of this macro
usages in cases where the type of the freed pointer is only partially
defined (e.g. only a forward declaration of it is available).