Compositor: Add Motion Vector type
The compositor previously overloaded the vector type to represent
multiple semantic types that are always stored in a 4D float vector. The
main types represented are Motion Vectors and 3D vectors. This patch
introduce a dedicated type for Motion Vector, leaving the vector type to
always represent a 3D vector, which will be done in a later commit.

This is not exposed to the user as a separate socket type with a
different color, but it is differentiated using the velocity subtype.