Refactor/Fix: WM messagebus: invalid usage of calloc/free on data embedding PointerRNA.
Add a dedicated callback to duplicate the key for each type.

RNA messagebus key is duplicated using move constructor, as memcpy is
not usable in these cases.

Keeping changes to the minimum necessary ones, the whole code and data
structure could use a serious update using C++ features, but this would
be totally out of scope of this fix.