Fix: Adding new objects in HIP-RT rendered viewport errors out
The issue was caused by c0ba800f64b.

Simple solution for now: check the data size and free the memory to
allow the device memory to be re-allocated. Seems the safest for the
upcoming release.

Ideally we'd need to avoid having these manual tricks with the device
memory pointers.