Fix nondeterminstic grid fill properly.
The previous behavior of edbm_grid_fill_prepare found the vertex with the largest angle, then walked the edge loop until one of the three best remaining corners was found.  This established one edge of the grid area, and the second edge was found by symmetry.

However, the direction that the algorithm walked the loop (and therefore *which* second corner was found) was essentially random -- the algorithm iterated in the direction of the vertex order of whichever selected edge happened to be the first one found during BM_ITER_MESH during BM_mesh_edgeloops_find. Since the dirction was random, sometimes edbm_grid_fill_prepare found the "row" and sometimes it found the "column".

The new behavior first finds the vert with the largest angle, then finds the vert with the second largest angle, (after excluding the diagonally opposite corner from the first vert), and selects the edge between those two verts

With this change, edbm_grid_fill_prepare always returns the same result for the same selection and same vertex coordinates, regardless of vertex order or edge iteration order.

This reverts commit 376e309b40f6739367137b0ce23faaa48b887c17 which somewhat masked this issue and is not necessary.